requirements.txt
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.0.2
requests==2.31.0
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.32.1
